requirements-dev.txt
cbor2==5.6.3
evdev==1.7.0
luma.core==2.4.2
luma.emulator==1.5.0
luma.oled==3.13.0
pillow==10.3.0
psutil==5.9.8
pyftdi==0.55.4
pygame==2.5.2
pynput==1.7.6
pyserial==3.5
python-xlib==0.33
pyusb==1.2.1
six==1.16.0
smbus2==0.4.3
spidev==3.6
